+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2024-04-17
+### Added
+- Introdotto meccanismo per attribuzione causale pausa pranzo durante la rispettiva fascia oraria 
+
+## [1.3.0] - 2023-03-14
+### Added
+- Corretto esempio Regex di Default
+- Introdotta CI per la build dell'immagine docker
+
 ## [1.2.3] - 2021-03-08
 ### Changed
 - Ritorno alla python:3-alpine con builder separato per diminuire dimensioni immagine,